Homeword

+++ Buổi 3/Homeword/BT3/cart.php
@@ -13,6 +13,17 @@ function tc_comment_out_shortcode( $amount) {
 	}
 }
 ?>
+
+<?php
+$cart = [
+[
+	'id' => '34567',
+	'name' => 'nokia 6300',
+	'amount' => '3',
+	'price' => '120002'
+]
+] ;
+?>
 <!DOCTYPE html>
 <html lang="en">
 <head>
@@ -38,17 +49,20 @@ function tc_comment_out_shortcode( $amount) {
 			</thead>
 			<tbody>
 				<?php 
-				foreach ($_SESSION['cart'] as $key => $value) {
-					?>
-					<tr>
-						<td><?php echo $value['id']; ?></td>
-						<td><?php echo $value['name']; ?></td>
-						<td><?php echo $value['amount']; ?></td>
-						<td><?php echo number_format($value['price']); ?></td>
-						<td><?php echo number_format($value['price']*$value['amount']); ?></td>
-						<td><a href="delete.php?id=<?php echo $value['id']; ?>" class="btn btn-danger">Xóa</a></td>
-					</tr>
-				<?php }
+				if (!isset($_SESSION['cart'])){
+					foreach ($_SESSION['cart'] as $key => $value) {
+						?>
+
+						<tr>
+							<td><?php echo $value['id']; ?></td>
+							<td><?php echo $value['name']; ?></td>
+							<td><?php echo $value['amount']; ?></td>
+							<td><?php echo number_format($value['price']); ?></td>
+							<td><?php echo number_format($value['price']*$value['amount']); ?></td>
+							<td><a href="delete.php?id=<?php echo $value['id']; ?>" class="btn btn-danger">Xóa</a></td>
+						</tr>
+					<?php }
+				}
 				
 				?>
 			</tbody>
